python/requirements.txt
docker-compose>=1.7
